Allow using instance IAM roles
Concourse do not allow using instance IAM roles because it conflicts
with their multi-tenancy designs[1]. Multiple "teams" can use the same
instance, and using IAM instance roles means that this could be
considered insecure.

This is not applicable to our use case, so we are fine to use it. This
removes the requirement to pass in access keys and just assume we want
to use instance roles instead.

[1] https://github.com/concourse/concourse/issues/3023